+++ lib/bones_api_db_directory.dart
@@ -1,0 +1,4 @@
+/// Bones_API DB Directory Adapters.
+library bones_api_adapter.db_directory;
+
+export 'src/bones_api_entity_db_object_directory.dart';