+++ test/bones_api_entity_db_tests_base.dart
@@ -623,7 +623,7 @@ Future<bool> runAdapterTests(
         expect(
           error.toString(),
           matches(RegExp(
-              r'Invalid entity\((?:User)?@table:user\) field(?:\(.*?email.*?\))?> reason: unique ; value: <.*?joe@\w+\.com.*?>.*',
+              r'Invalid entity\((?:User)?@table:user\) field(?:\(.*?email.*?\))?> reason: unique ; value: <.*?joe@[\w.+]+\.com.*?>.*',
               dotAll: true)),
         );
       }